Convert static methods to utils fxs
